[12.x] New `illuminate/reflections` component from `illuminate/support`
