+++ composer.json
@@ -90,6 +90,7 @@
         "illuminate/pipeline": "self.version",
         "illuminate/process": "self.version",
         "illuminate/queue": "self.version",
+        "illuminate/reflections": "self.version",
         "illuminate/redis": "self.version",
         "illuminate/routing": "self.version",
         "illuminate/session": "self.version",
@@ -143,6 +144,7 @@
             "src/Illuminate/Filesystem/functions.php",
             "src/Illuminate/Foundation/helpers.php",
             "src/Illuminate/Log/functions.php",
+            "src/Illuminate/Reflections/helpers.php",
             "src/Illuminate/Support/functions.php",
             "src/Illuminate/Support/helpers.php"
         ],
@@ -151,7 +153,8 @@
             "Illuminate\\Support\\": [
                 "src/Illuminate/Macroable/",
                 "src/Illuminate/Collections/",
-                "src/Illuminate/Conditionable/"
+                "src/Illuminate/Conditionable/",
+                "src/Illuminate/Reflections/"
             ]
         }
     },
