+++ src/Illuminate/Container/composer.json
@@ -16,6 +16,7 @@
     "require": {
         "php": "^8.2",
         "illuminate/contracts": "^12.0",
+        "illuminate/reflections": "^12.0",
         "psr/container": "^1.1.1|^2.0.1",
         "symfony/polyfill-php84": "^1.33",
         "symfony/polyfill-php85": "^1.33"
