+++ src/Illuminate/Reflections/Traits/ReflectsClosures.php
