+++ src/Illuminate/Support/composer.json
@@ -23,6 +23,7 @@
         "illuminate/conditionable": "^12.0",
         "illuminate/contracts": "^12.0",
         "illuminate/macroable": "^12.0",
+        "illuminate/reflections": "^12.0",
         "nesbot/carbon": "^3.8.4",
         "symfony/polyfill-php83": "^1.33",
         "symfony/polyfill-php85": "^1.33",
